refactor: storage 구조 정리
