+++ apps/backend/src/modules/users/users.repository.ts
@@ -5,7 +5,7 @@ import {
   NotFoundException,
 } from "@nestjs/common";
 // biome-ignore lint/style/useImportType: Nest DI needs a runtime class reference.
-import { SupabaseService } from "../storage/supabase/supabase.service";
+import { SupabaseService } from "../../storage/supabase/supabase.service";
 // biome-ignore lint/style/useImportType: Nest DI needs a runtime class reference.
 import { CreateGuestUserInput } from "./dto/create-guest-user.dto";
 // biome-ignore lint/style/useImportType: Nest DI needs a runtime class reference.
